tests/wpt/web-platform-tests/referrer-policy/same-origin/http-rp/cross-origin/http-http/script-tag/no-redirect/cross-origin.http.html

<!DOCTYPE html>
<!-- DO NOT EDIT! Generated by referrer-policy/generic/tools/generate.py using referrer-policy/generic/template/test.release.html.template. -->
<html>
  <head>
    <title>Referrer-Policy: Referrer Policy is set to 'same-origin'</title>
    <meta name="description" content="Check that cross-origin subresources get no referrer information and same-origin get the stripped referrer URL.">
    <!-- No meta: Referrer policy delivered via HTTP headers. -->
    <link rel="author" title="Kristijan Burnik" href="burnik@chromium.org">
    <link rel="help" href="https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-same-origin">
    <meta name="assert" content="The referrer URL is omitted when a
                                 document served over http requires an http
                                 sub-resource via script-tag using the http-rp
                                 delivery method with no-redirect and when
                                 the target request is cross-origin.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/security-features/resources/common.js"></script>
    <script src="/referrer-policy/generic/referrer-policy-test-case.js?pipe=sub"></script>
  </head>
  <body>
    <script>
      ReferrerPolicyTestCase(
        {
          "referrer_policy": "same-origin",
          "delivery_method": "http-rp",
          "redirection": "no-redirect",
          "origin": "cross-origin",
          "source_protocol": "http",
          "target_protocol": "http",
          "subresource": "script-tag",
          "referrer_url": "omitted"
        },
        document.querySelector("meta[name=assert]").content,
        new SanityChecker()
      ).start();
      </script>
    <div id="log"></div>
  </body>
</html>
