LayoutTests/imported/w3c/web-platform-tests/referrer-policy/same-origin/meta-referrer/cross-origin/http-http/script-tag/no-redirect/cross-origin.http-expected.txt


PASS The referrer URL is omitted when a
                                 document served over http requires an http
                                 sub-resource via script-tag using the meta-referrer
                                 delivery method with no-redirect and when
                                 the target request is cross-origin.
